Add MediaRepository and AppContainer
The repository holds the list of available media (for now in memory) and
is created in the AppContainer that lives in the application class.
Through the AppContainer the rest of the app can access the repository
and be able to find what media is currently played with just the id.