Move NetworkClient to libsoli
The network client and its API is common to the android and ios client
app but they both have to provide a valid engine for their respective
platform.
The models are shared as they are.
The PodcastFeedParser need to be implemented per platform because the
lib used to parse the xml is not yet multiplatform.